fix editor usage of graphic controller...
The old code would force the objects to maintain references to textures stored in asset bundles, even while in the editor. Technically it worked, but Unity hates it, and the GUID errors it generates are now causing Windows builds to fail. So, this commit fixes the problem by keeping a direct reference to an editor sprite. Related improvements have been made as well.